added support for removing inside a topic, and refactored some repeated code into methods

+++ server/src/controllers/MemeSoundController.ts
@@ -1,7 +1,7 @@
 import {getRepository} from "typeorm";
 import {NextFunction, Request, Response} from "express";
 import { MemeSound } from "../entity/MemeSound";
-import { getFromTableRandom } from "./MemeControllerHelperMethods";
+import { getElement, getFromTableRandom, getTopic, isModerator, verifyUser } from "./MemeControllerHelperMethods";
 import * as url from 'url';
 import {soundsFolder, mediaHost, uploadfolder} from '../index'
 import { Topic } from "../entity/Topic";
@@ -34,12 +34,8 @@ export class MemeSoundController {
         let sound = new MemeSound();
         if(req.params.topic){
             
-            let topic;
-
-            try{
-                topic = getRepository(Topic).findOneOrFail({where: {name: req.params.topic}})
-            } catch (error){
-                res.status(404).send({error: "Topic not found"});
+            let topic = await getTopic(req,res);
+            if (!topic){
                 return;
             }
             sound.topic = topic
@@ -51,14 +47,51 @@ export class MemeSoundController {
         return this.memeSoundRepository.save(sound);
     }
 
+    
     async remove(req: Request, res: Response, next: NextFunction) {
+
+        let id = req.params.id
+
+        if(!id){
+            res.status(400).send("Bad request");
+            return;
+        }
+
+        if(req.params.topic){
+
+            let topic = await getTopic(req,res,['moderators','owner']);
+            if (!topic){
+                return;
+            }
+
+            let user = await verifyUser(res);
+            if(!user){
+                return;
+            }
+
+            if(!isModerator(topic,user) && user.role != 'ADMIN'){
+                res.status(401).send({error: "User is not a moderator of this topic"});
+                return;
+            }
+
+            let soundToRemove = await getElement(MemeSound,req,res,"Toptext not found in topic",topic);
+            if(!soundToRemove){
+                return;
+            }
+
+            return this.memeSoundRepository.remove(soundToRemove)
+        }
+
         if (req.body.SECRET !== process.env.SECRET){
             res.status(403);
             return {you:"suck"};
         }
-        
-        const soundToRemove = await this.memeSoundRepository.findOne(req.params.id);
-        return await this.memeSoundRepository.remove(soundToRemove);
+        let soundToRemove = await getElement(MemeSound,req,res,"Toptext not found in topic");
+        if(!soundToRemove){
+            return;
+        }
+
+        return this.memeSoundRepository.remove(soundToRemove);
     }
 
     async random(req: Request, res: Response, next: NextFunction) {
