+++ server/src/controllers/MemeVisualController.ts
@@ -1,7 +1,7 @@
 import {getRepository} from "typeorm";
 import {NextFunction, Request, Response} from "express";
 import { MemeVisual } from "../entity/MemeVisual";
-import { getFromTableRandom, saveVerifyCompress } from "./MemeControllerHelperMethods";
+import { getElement, getFromTableRandom, getTopic, isModerator, saveVerifyCompress, verifyUser } from "./MemeControllerHelperMethods";
 import * as url from "url";
 import {mediaHost, uploadfolder, visualsFolder} from '../index'
 
@@ -35,12 +35,8 @@ export class MemeVisualController {
         let visual = new MemeVisual();
         if(req.params.topic){
             
-            let topic;
-
-            try{
-                topic = getRepository(Topic).findOneOrFail({where: {name: req.params.topic}})
-            } catch (error){
-                res.status(404).send({error: "Topic not found"});
+            let topic = await getTopic(req,res);
+            if (!topic){
                 return;
             }
             visual.topic = topic
@@ -56,13 +52,51 @@ export class MemeVisualController {
     }
 
     async remove(req: Request, res: Response, next: NextFunction) {
+
+        let id = req.params.id
+
+        if(!id){
+            res.status(400).send("Bad request");
+            return;
+        }
+
+        if(req.params.topic){
+
+            let topic = await getTopic(req,res,['moderators','owner']);
+            if (!topic){
+                return;
+            }
+
+            let user = await verifyUser(res);
+            if(!user){
+                return;
+            }
+
+            if(!isModerator(topic,user) && user.role != 'ADMIN'){
+                res.status(401).send({error: "User is not a moderator of this topic"});
+                return;
+            }
+
+            let visualToRemove = await getElement(MemeVisual,req,res,"Toptext not found in topic",topic);
+            if(!visualToRemove){
+                return;
+            }
+
+            return this.memeVisualRepository.remove(visualToRemove)
+        }
+
         if (req.body.SECRET !== process.env.SECRET){
             res.status(403);
             return {you:"suck"};
         }
 
-        const visualToRemove = await this.memeVisualRepository.findOne(req.params.id);
-        return await this.memeVisualRepository.remove(visualToRemove);
+        let visualToRemove = await getElement(MemeVisual,req,res,"Toptext not found in topic");
+        if(!visualToRemove){
+            return;
+        }
+
+        return this.memeVisualRepository.remove(visualToRemove);
+
     }
 
     async random(req: Request, res: Response, next: NextFunction) {
